backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
llama-cpp-python==0.2.38